+++ lib/workers/branch.js
@@ -1,13 +1,16 @@
-const logger = require('../helpers/logger');
 const handlebars = require('handlebars');
 const packageJsonHelper = require('../helpers/package-json');
 const npmHelper = require('../helpers/npm');
 const yarnHelper = require('../helpers/yarn');
+const prWorker = require('./pr');
 
 module.exports = {
   getParentBranch,
   ensureBranch,
+  updateBranch,
 };
+
+let logger;
 
 async function getParentBranch(branchName, config) {
   // Check if branch exists
@@ -187,4 +190,49 @@ async function ensureBranch(upgrades) {
   }
   // Return true as branch exists
   return true;
+}
+
+async function updateBranch(upgrades, parentLogger) {
+  const upgrade0 = upgrades[0];
+  // Use templates to generate strings
+  const branchName = handlebars.compile(upgrade0.branchName)(upgrade0);
+  const prTitle = handlebars.compile(upgrade0.prTitle)(upgrade0);
+
+  logger = parentLogger.child({
+    repository: upgrade0.repository,
+    branch: branchName,
+  });
+
+  const upgradeCount = upgrades.length === 1
+    ? '1 upgrade'
+    : `${upgrades.length} upgrades`;
+  logger.info(
+    `Branch '${branchName}' has ${upgradeCount}: ${upgrades.map(
+      upgrade => upgrade.depName
+    )}`
+  );
+
+  try {
+    if (
+      upgrade0.upgradeType !== 'maintainYarnLock' &&
+      upgrade0.groupName === null &&
+      !upgrade0.recreateClosed &&
+      (await upgrade0.api.checkForClosedPr(branchName, prTitle))
+    ) {
+      logger.info(
+        `Skipping ${branchName} upgrade as matching closed PR already existed`
+      );
+      return;
+    }
+    const branchCreated = await module.exports.ensureBranch(upgrades);
+    if (branchCreated) {
+      const pr = await prWorker.ensurePr(upgrades, logger);
+      if (pr) {
+        await prWorker.checkAutoMerge(pr, upgrade0, logger);
+      }
+    }
+  } catch (error) {
+    logger.error(`Error updating branch ${branchName}: ${error}`);
+    // Don't throw here - we don't want to stop the other renovations
+  }
 }
