+++ lib/workers/package-file.js
@@ -4,7 +4,6 @@ const npmApi = require('../api/npm');
 const packageJson = require('../helpers/package-json');
 const versionsHelper = require('../helpers/versions');
 
-let config;
 let logger;
 
 module.exports = {
@@ -15,15 +14,16 @@ module.exports = {
 };
 
 // This function manages the queue per-package file
-async function processPackageFile(repository, packageFile, packageConfig) {
-  // Initialize globals
-  config = Object.assign({}, packageConfig);
-  config.packageFile = packageFile;
-  logger = config.logger.child({ repository, packageFile }, true);
+async function processPackageFile(config) {
+  // Initialize logger
+  logger = config.logger.child({
+    repository: config.repository,
+    packageFile: config.packageFile,
+  });
 
-  logger.info(`Processing package file ${repository}:${packageFile}`);
+  logger.info(`Processing package file`);
 
-  const packageContent = await config.api.getFileJson(packageFile);
+  const packageContent = await config.api.getFileJson(config.packageFile);
 
   if (!packageContent) {
     logger.warn('No package.json content found - skipping');
@@ -106,6 +106,7 @@ function assignDepConfigs(inputConfig, deps) {
         }
       });
     }
+    // TODO: clean this up
     delete returnDep.config.depType;
     delete returnDep.config.depTypes;
     delete returnDep.config.enabled;
