+++ lib/workers/pr.js
@@ -1,4 +1,3 @@
-const logger = require('../helpers/logger');
 const handlebars = require('handlebars');
 const changelogHelper = require('../helpers/changelog');
 
@@ -8,7 +7,7 @@ module.exports = {
 };
 
 // Ensures that PR exists with matching title/body
-async function ensurePr(upgrades) {
+async function ensurePr(upgrades, logger) {
   logger.debug({ config: upgrades }, 'ensurePr');
   // If there is a group, it will use the config of the first upgrade in the array
   const config = Object.assign({}, upgrades[0]);
@@ -140,7 +139,7 @@ async function ensurePr(upgrades) {
   return null;
 }
 
-async function checkAutoMerge(pr, config) {
+async function checkAutoMerge(pr, config, logger) {
   logger.debug(`Checking #${pr.number} for automerge`);
   if (config.automergeEnabled && config.automergeType === 'pr') {
     logger.info('PR is configured for automerge');
