+++ lib/workers/repository.js
@@ -1,14 +1,21 @@
+// Global requires
+const handlebars = require('handlebars');
+const ini = require('ini');
+const stringify = require('json-stringify-pretty-compact');
+// API
 const githubApi = require('../api/github');
 const gitlabApi = require('../api/gitlab');
 const npmApi = require('../api/npm');
+// Config
 const defaultsParser = require('../config/defaults');
-const ini = require('ini');
-const packagesWorker = require('./packages');
+// Workers
 const packageFileWorker = require('./package-file');
-const stringify = require('json-stringify-pretty-compact');
+const branchWorker = require('./branch');
 
 module.exports = {
   processRepo,
+  processUpgrades,
+  removeStandaloneBranches,
 };
 
 // This will be github or others
@@ -18,6 +25,7 @@ let logger;
 // Queue package files in sequence within a repo
 async function processRepo(config) {
   logger = config.logger.child({ repository: config.repository });
+  config.logger = logger; // eslint-disable-line no-param-reassign
   logger.info('Renovating repository');
   logger.debug({ config }, 'processRepo');
   if (config.platform === 'github') {
@@ -46,7 +54,7 @@ async function processRepo(config) {
     // Detect package files if none already configured
     await detectPackageFiles(config);
     const upgrades = await getAllRepoUpgrades(config);
-    await packagesWorker.processUpgrades(upgrades);
+    await module.exports.processUpgrades(upgrades);
   } catch (error) {
     throw error;
   }
@@ -164,15 +172,80 @@ async function getAllRepoUpgrades(config) {
     }
     const cascadedConfig = Object.assign({}, config, packageFile);
     // Remove unnecessary fields
-    delete cascadedConfig.repository;
+    cascadedConfig.packageFile = cascadedConfig.fileName;
     delete cascadedConfig.fileName;
     upgrades = upgrades.concat(
-      await packageFileWorker.processPackageFile(
-        config.repository,
-        packageFile.fileName,
-        cascadedConfig
-      )
+      await packageFileWorker.processPackageFile(cascadedConfig)
     );
   }
   return upgrades;
+}
+
+async function processUpgrades(upgrades) {
+  if (upgrades.length) {
+    const upgradeCount = upgrades.length === 1
+      ? '1 dependency upgrade'
+      : `${upgrades.length} dependency upgrades`;
+    logger.info(`Processing ${upgradeCount}`);
+  } else {
+    logger.info('No upgrades to process');
+  }
+  logger.debug({ config: upgrades }, 'All upgrades');
+  const branchUpgrades = {};
+  for (const upgrade of upgrades) {
+    const flattened = Object.assign({}, upgrade.config, upgrade);
+    delete flattened.config;
+    if (flattened.upgradeType === 'pin') {
+      flattened.isPin = true;
+    } else if (flattened.upgradeType === 'major') {
+      flattened.isMajor = true;
+    } else if (flattened.upgradeType === 'minor') {
+      flattened.isMinor = true;
+    }
+    // Check whether to use a group name
+    let branchName;
+    if (flattened.groupName) {
+      logger.debug(
+        `Dependency ${flattened.depName} is part of group '${flattened.groupName}'`
+      );
+      flattened.groupSlug =
+        flattened.groupSlug ||
+        flattened.groupName.toLowerCase().replace(/[^a-z0-9+]+/g, '-');
+      branchName = handlebars.compile(flattened.groupBranchName)(flattened);
+      logger.debug(`branchName=${branchName}`);
+      if (branchUpgrades[branchName]) {
+        // flattened.branchName = flattened.groupBranchName;
+        flattened.commitMessage = flattened.groupCommitMessage;
+        flattened.prTitle = flattened.groupPrTitle;
+        flattened.prBody = flattened.groupPrBody;
+      }
+    } else {
+      branchName = handlebars.compile(flattened.branchName)(flattened);
+    }
+    branchUpgrades[branchName] = branchUpgrades[branchName] || [];
+    branchUpgrades[branchName] = [flattened].concat(branchUpgrades[branchName]);
+  }
+  logger.debug({ config: branchUpgrades }, 'Branched upgrades');
+  for (const branch of Object.keys(branchUpgrades)) {
+    await module.exports.removeStandaloneBranches(branchUpgrades[branch]);
+    await branchWorker.updateBranch(branchUpgrades[branch], logger);
+  }
+}
+
+async function removeStandaloneBranches(upgrades) {
+  if (upgrades.length > 1) {
+    for (const upgrade of upgrades) {
+      const standaloneBranchName = handlebars.compile(upgrade.branchName)(
+        upgrade
+      );
+      logger.debug(`Need to delete branch ${standaloneBranchName}`);
+      try {
+        await upgrade.api.deleteBranch(standaloneBranchName);
+      } catch (err) {
+        logger.debug(`Couldn't delete branch ${standaloneBranchName}`);
+      }
+      // Rename to group branchName
+      upgrade.branchName = upgrade.groupBranchName;
+    }
+  }
 }
